hotfix MPVERKEHR-186 fhh-atlas Gurlittinsel
The first week after a leap year has the wrong value in the GFI
for Gurlittinsel. (e.g. 2021: KW52 of 2020 instead of KW53 of 2020).

A dataset from the Gurlittinsel like "2021,-53,0" means: KW53 of 2020.
moment.js does not work by setting a week to -53, as it will substract
53 weeks (landing in KW52 of 2019).

Now we decrement the year if a negative week number is given.

Also fixed:
- if the resulting date is invalid (e.g. KW53 of 2021), the data is ignored
- we now use UTC time, so the start of the week in Hamburg is always sunday,
23:00:00 London Time at winter and 22:00:00 London Time at summer