hotfix MPVERKEHR-186 fhh-atlas Gurlittinsel gaps
GFI theme Gurlittinsel: The algorithm calculating the gaps
between labels of the x axis isn't up to date anymore.

I changed the algorithm using a threshold to increment gaps
by 1 for the number of thresholds fitting into data.length.

Adjusted the threshold for day, week and year as follows:
- day: every second hour is shown
- week: every day has a label
- year: for up to 20KW every kw, until 40KW every second kw and
for up to 60KW every third KW is shown as label

Also fixed css issues with gfi height and text alignment.