G31DEV1-90 Fix: no data in gfi theme of verkehrs layers in geo-online
In geo online if we active the layer of verkehr and click them to open
a gfi theme, there will be no data.

The reason is that the mqtt configuration is updated, but for verkehr's
Layer, the mqtt configuration in gfi theme trafficCount is not updated.
After updating the configuration options, it works again.